Creates a twitter analysis pipeline
Despite that the particular repo and PR is focused
on twitter, they are not only about it.
We have implemented yaml. kafka, faust over python
offers a great way to go when it comes to pocess
inoformtion feed. Keras and TF should work fine next
to such a solution.

This is a squashed commit.